requirements.txt
pdfplumber
python-docx
lxml